Divided replacer and had to make another file
