+++ src/replacer/replacer.c
@@ -6,7 +6,7 @@
 /*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 17:55:44 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/11/06 14:09:00 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/11/06 14:21:11 by ddiniz-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,55 +14,6 @@
 
 char	**var_split(char *str);
 char	*var_iter(t_list **env, char *var);
-
-char	*replace_dollar(char *str, char *buf, t_list **env, int flag)
-{
-	char	*res;
-	char	*buf1;
-	char	*buf2;
-
-	buf2 = NULL;
-	if (str[0] && str[1] && str[0] == '$' && str[1] == '?')
-		return (dollar_cond(buf));
-	if (ft_strcmp(str, "$") == 0 || (str[1] && !ft_isalnum(str[1])
-			&& str[1] != '_' && str[1] != '\\') || str[1] == '?')
-		return (ft_strjoin(buf, str));
-	buf1 = ft_strtrim(str, "$");
-	buf2 = replace_str(buf1, env);
-	free(buf1);
-	if (flag == 2)
-	{
-		buf1 = ft_strdup(buf2);
-		free(buf2);
-		buf2 = add_quotes(buf1, '\'');
-		free(buf1);
-	}
-	res = ft_strjoin(buf, buf2);
-	free(buf2);
-	return (res);
-}
-
-char	*replace_quotes(char *str, char *buf, t_list **env, int flag)
-{
-	char	*res;
-	char	*buf1;
-
-	res = NULL;
-	buf1 = NULL;
-	if (str[0] == '\'' && ft_strcmp(str, "\'\'") != 0)
-		res = replace_single(str, buf, env, flag);
-	else if (str[0] == '\"' && ft_strcmp(str, "\"\"") != 0)
-	{
-		res = remove_quotes(str, '\"');
-		buf1 = replacer(res, env, 1);
-		free(res);
-		res = ft_strjoin(buf, buf1);
-	}
-	else
-		res = ft_strjoin(buf, str);
-	free(buf1);
-	return (res);
-}
 
 char	*replace_cond(char *str, char *buf1, t_list **env, int flag)
 {
@@ -86,19 +37,14 @@ char	*replace_cond(char *str, char *buf1, t_list **env, int flag)
 	return (res);
 }
 
-//Joins all substituted strings from the split str
-char	*replacer(char *str, t_list **env, int flag)
+char	*replace_loop(char **arr, t_list **env, int flag)
 {
 	int		i;
 	char	*res;
 	char	*buf1;
-	char	**arr;
 
 	i = 0;
 	res = NULL;
-	arr = var_split(str);
-	if (!arr)
-		return (NULL);
 	while (arr[i])
 	{
 		buf1 = ft_strdup(res);
@@ -114,6 +60,21 @@ char	*replacer(char *str, t_list **env, int flag)
 		free(buf1);
 		i++;
 	}
+	return (res);
+}
+
+//Joins all substituted strings from the split str
+char	*replacer(char *str, t_list **env, int flag)
+{
+	int		i;
+	char	*res;
+	char	**arr;
+
+	i = 0;
+	arr = var_split(str);
+	if (!arr)
+		return (NULL);
+	res = replace_loop(arr, env, flag);
 	free_array(arr);
 	return (res);
 }
